requirements.txt
fastapi
uvicorn[standard]==0.24.0
sqlalchemy
psycopg2-binary==2.9.9
pydantic
python-dotenv==1.0.0
typing-extensions==4.8.0
email-validator==2.1.0.post1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.18
alembic
pytest==7.4.3
httpx==0.25.1
pandas~=2.2.3
PyJWT